Add PubgReportHostedService class
The new class PubgReportHostedService has been added to deal with the hosting services. This class is designed to start and stop a timer asynchronously, and fetch stream information for a specified account during each iteration. The implementation for sending data to Discord is marked as a future task and currently not implemented.