Add NUnit test suite for PubgReportCrawler
This commit includes the addition of a test project for PubgReportCrawler using the NUnit framework. It includes creation of new StreamInfoServiceTests class with several test cases, setup of the test project in the solution file, addition of NUnit as a global using, and package references related to NUnit into the PubgReportCrawlerTests.csproj file.